refactor to global functions
