+++ packages/react-typescript/package.json
@@ -10,6 +10,15 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@bit/bit.envs.compilers.typescript": "^3.1.1"
+    "@bit/bit.envs.common.build-component": "0.0.1",
+    "@bit/bit.envs.compilers.typescript": "^3.1.1",
+    "@types/fs-extra": "^8.0.1",
+    "@types/rimraf": "^2.0.3",
+    "fs-extra": "^8.1.0",
+    "rimraf": "^3.0.0"
+  },
+  "devDependencies": {
+    "@types/chai": "^4.2.5",
+    "chai": "^4.2.0"
   }
 }
