+++ packages/react-typescript/src/custom-types.ts
@@ -1,0 +1,64 @@
+export function getCustomTypes() {
+  return `
+      declare module '*.bmp' {
+        const src: string;
+        export default src;
+      }
+      
+      declare module '*.gif' {
+        const src: string;
+        export default src;
+      }
+      
+      declare module '*.jpg' {
+        const src: string;
+        export default src;
+      }
+      
+      declare module '*.jpeg' {
+        const src: string;
+        export default src;
+      }
+      
+      declare module '*.png' {
+        const src: string;
+        export default src;
+      }
+      
+      declare module '*.webp' {
+          const src: string;
+          export default src;
+      }
+      
+      declare module '*.ico' {
+        const image: string;
+        export default image;
+      }
+      
+      
+      declare module '*.svg' {
+        import * as React from 'react';
+      
+        export const ReactComponent: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+      
+        const src: string;
+        export default src;
+      }
+      
+      
+      declare module '*.css' {
+        const classes: { readonly [key: string]: string };
+        export default classes;
+      }
+      
+      declare module '*.scss' {
+        const classes: { readonly [key: string]: string };
+        export default classes;
+      }
+      
+      declare module '*.sass' {
+        const classes: { readonly [key: string]: string };
+        export default classes;
+      }    
+    `;
+}
