--- packages/ts-compiler/test/typescript-react.ts
@@ -1,4 +1,0 @@
-import { TypescriptCompiler } from '../src/typescript-compiler';
-import { presetStore } from '../src/preset';
-
-export default  new TypescriptCompiler(presetStore['REACT']);
